requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
openai
faiss-cpu
numpy
pandas
streamlit
streamlit-pdf-viewer
requests
python-dotenv
email-validator
PyPDF2
python-docx
docx
tools
pytesseract
Pillow
pdf2image
groq
poppler-utils